poldit-ai-docker/flask_app/requirements.txt
amqp==5.1.1
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
prompt-toolkit==3.0.39
pycparser==2.21
PyJWT==2.7.0
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.31.0
requests-file==1.5.1
six==1.16.0
tldextract==3.4.4
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
zope.event==5.0
zope.interface==6.0
gunicorn==20.1.0